Build fix

+++ tests/PopulationTests.hs
@@ -51,7 +51,7 @@ prop_unlikely_evidence =
 
 -- Perfectly distributed random numbers between 0 and 1. In a particular random order (sorted!) :D
 many_tosses :: Fractional a => Int -> [a]
-many_tosses how_many = take how_many $ map (\n -> (fromIntegral n) / (fromIntegral how_many)) $ [0..]
+many_tosses how_many = take how_many $ map (\n -> (fromIntegral n) / (fromIntegral how_many)) $ [0::Int ..] 
 
 -- Populations. A causal system is a generator of populations. The input is generators of facts.
 -- The output is population of facts.
